+++ examples/node/cloud.js
@@ -1,4 +1,4 @@
-module.exports = async function(notion) {
-  const info = await notion.getInfo();
+module.exports = async function (neurosity) {
+  const info = await neurosity.getInfo();
   console.log("info", info);
 };