+++ examples/node/kinesis.js
@@ -1,5 +1,5 @@
-module.exports = async function(notion) {
-  notion.kinesis("leftHandPinch").subscribe(kinesis => {
+module.exports = async function (neurosity) {
+  neurosity.kinesis("leftHandPinch").subscribe((kinesis) => {
     console.log("kinesis", kinesis);
   });
 };