+++ examples/node/rawUnfiltered.js
@@ -1,6 +1,6 @@
-module.exports = async function (notion) {
+module.exports = async function (neurosity) {
   // data does not get notch or band pass filters applied unlike the `raw` option
-  notion.brainwaves("rawUnfiltered").subscribe((brainwaves) => {
+  neurosity.brainwaves("rawUnfiltered").subscribe((brainwaves) => {
     console.log("brainwaves", brainwaves);
   });
 };