+++ examples/node/subscriptions.js
@@ -1,42 +1,42 @@
-const { Notion } = require("../..");
+const { Neurosity } = require("../..");
 
-const notion = new Notion({
+const neurosity = new Neurosity({
   deviceId: process.env.NEUROSITY_DEVICE_ID
 });
 
-const notion2 = new Notion({
+const neurosity2 = new Neurosity({
   deviceId: process.env.NEUROSITY_DEVICE_ID
 });
 
 (async () => {
-  await notion.login({
+  await neurosity.login({
     email: process.env.NEUROSITY_EMAIL,
     password: process.env.NEUROSITY_PASSWORD
   });
 
-  await notion2.login({
+  await neurosity2.login({
     email: process.env.NEUROSITY_EMAIL,
     password: process.env.NEUROSITY_PASSWORD
   });
 
-  const sub1 = notion.signalQuality().subscribe();
-  const sub2 = notion.calm().subscribe();
-  const sub3 = notion.focus().subscribe();
-  const sub4 = notion2.focus().subscribe();
+  const sub1 = neurosity.signalQuality().subscribe();
+  const sub2 = neurosity.calm().subscribe();
+  const sub3 = neurosity.focus().subscribe();
+  const sub4 = neurosity2.focus().subscribe();
 
   setTimeout(() => {
     sub2.unsubscribe();
   }, 2000);
 
   setTimeout(() => {
-    notion.goOffline();
+    neurosity.goOffline();
     setTimeout(() => {
       sub1.unsubscribe();
     }, 100);
   }, 6000);
 
   setTimeout(() => {
-    notion.goOnline();
+    neurosity.goOnline();
 
     setTimeout(() => {
       sub3.unsubscribe();