+++ src/types/signalQuality.ts
@@ -4,7 +4,7 @@ type ChannelQuality = {
 };
 
 /**
- * Channel names for Notion are `CP6, F6, C4, CP4, CP3, F5, C3 and CP5`
+ * Channel names for the Crown are `CP3, C3, F5, PO3, PO4, F6, C4, and CP4`
  * These channels may change in future versions
  * {@link ChannelQuality}
  */