+++ src/utils/errors.ts
@@ -1,6 +1,6 @@
 export const prefix = "Neurosity SDK: ";
 export const mustSelectDevice = new Error(
-  `${prefix}A device must be selected. Make sure to call "notion.selectDevice()"`
+  `${prefix}A device must be selected. Make sure to call "neurosity.selectDevice()"`
 );
 
 export const metricNotSupportedByModel = (
@@ -12,10 +12,7 @@ export const metricNotSupportedByModel = (
   );
 };
 
-export const locationNotFound = (
-  location: string,
-  modelVersion: string
-) => {
+export const locationNotFound = (location: string, modelVersion: string) => {
   return new Error(
     `${prefix}${location} location not supported on model version ${modelVersion}. Check spelling or see docs.neurosity.co for more info.`
   );